change /home to /
