+++ pages/settings.tsx
@@ -1,7 +1,20 @@
+import { signOut, useSession } from 'next-auth/react';
+import styles from '@/styles/Home.module.css';
 import React from 'react';
 
 const SettingsPage = () => {
-  return <>Settings</>;
+  const { data: session } = useSession();
+
+  return (
+    <div className={styles.container}>
+      {session && (
+        <>
+          <div>You have signed in as {session.user?.email}</div>
+          <button onClick={() => signOut()}>Sign out</button>
+        </>
+      )}
+    </div>
+  );
 };
 
 export default SettingsPage;
